+++ packages/dso-toolkit/src/components/image-overlay/image-overlay.models.ts
@@ -2,4 +2,5 @@ import { Image } from "../image/image.models.js";
 
 export interface ImageOverlay {
   image: Image;
+  title: string;
 }